Add fiches_de_poste_par_candidature table Also: (squashed to simplify merge conflict resolution) Move dataframe methods to shared library Move populate_departments to correct script Refactor get_df_from_rows method
